pyproject.toml
[project]
name = "recipe-app-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "psycopg[c]>=3.2.8",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
]
type = [
    "django-stubs>=5.2.0",
    "djangorestframework-stubs>=3.16.0",
]
[tool.ruff]
exclude = [
    "**/migrations/*.py",
    ".venv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
ignore = ["ISC001", "D203", "D213", "D104"]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PTH", # flake8-use-pathlib
    "PD", # pandas-vet
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"**/migrations/**.py" = ["D100", "D101", "D102", "D103", "E501", "RUF012"]
"**/tests/**/*.py" = ["S101", "E501", "PLR0917", "ISC001", "D100", "D101", "D102", "PLR0913", "PLR2004"]
"**/*serializer*.py" = ["N815", "RUF012"]
"**/views.py" = ["N815", "RUF012"]
"**/__init__.py" = ["D104"]
"**/models.py" = ["COM819"]
"**/admin.py" = ["RUF012"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
